requirements.txt
beautifulsoup4
lxml
sentence-transformers
faiss-cpu
chromadb
numpy
requests